LayoutTests/fast/exclusions/parsing/parsing-wrap-through-expected.txt

Test parsing of the CSS wrap-through property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getCSSText("-webkit-wrap-through", "wrap") is "wrap"
PASS getCSSText("-webkit-wrap-through", "none") is "none"
PASS getCSSText("-webkit-wrap-through", ";") is ""
PASS getCSSText("-webkit-wrap-through", "5") is ""
PASS getCSSText("-webkit-wrap-through", "-1.2") is ""
PASS getCSSText("-webkit-wrap-through", "'string'") is ""
PASS getComputedStyleValue("-webkit-wrap-through", "wrap") is "wrap"
PASS getComputedStyleValue("-webkit-wrap-through", "5") is "wrap"
PASS getComputedStyleValue("-webkit-wrap-through", "'string'") is "wrap"
PASS getChildComputedStyle("-webkit-wrap-through", "wrap", "none") is "none"
PASS getChildComputedStyle("-webkit-wrap-through", "none", "wrap") is "wrap"
PASS successfullyParsed is true

TEST COMPLETE
